Send structure and errors as FlowGenerationData and rename model files
