+++ frank-flow/src/frontend/src/app/flow/node/nodes/exit.model.ts
@@ -1,5 +1,5 @@
 import { Node } from './node.model';
-import { FlowNodeAttributes } from '../../../shared/models/flowNodeAttributes.model';
+import { FlowNodeAttributes } from '../../../shared/models/flow-node-attributes.model';
 
 export default class Exit extends Node {
   constructor(
