+++ frank-flow/src/frontend/src/app/shared/models/flow-generation-data.model.ts
@@ -1,0 +1,6 @@
+import { FlowStructure } from './flow-structure.model';
+
+export interface FlowGenerationData {
+  structure: FlowStructure;
+  errors: string[];
+}
