+++ frank-flow/src/frontend/src/app/shared/models/flow-node-attribute-options.model.ts
