+++ frank-flow/src/frontend/src/app/shared/models/flow-node-attribute.model.ts
