+++ frank-flow/src/frontend/src/app/shared/models/flow-node-attributes.model.ts
@@ -1,4 +1,4 @@
-import { FlowNodeAttribute } from './flowNodeAttribute.model';
+import { FlowNodeAttribute } from './flow-node-attribute.model';
 
 export interface FlowNodeAttributes {
   [p: string]: FlowNodeAttribute;
