+++ frank-flow/src/frontend/src/app/shared/models/flow-tree-node.model.ts
